Use ETS as a backend for mocked module function calls
Summary of changes introduced by this commit:

* Mock module call state in stored in ETS. Each module keeps a reference
to a unique anonymous ETS table, which is destroyed when its owner (the
test process) dies.
* A mock module can be defined in a `setup_all` keeping state between test cases.
* The :defmock application no longer needs to be started before ExUnit
* Mocks modules are named uniquely based on the context of their definition
* called_with? is changed to have arity of 2 (instead of accepting any number of arguments)
to conform with the Erlang / Elixir convention where variable arity is very rare.
* called_with? has changed to not be agnostic of the order of Keyword Lists, since order their matters
in pattern-matching. This can still be achieved using a valid match_spec expression in `called_with_match?` (see below).
* called_with_match? is added, accepting a [match_spec][match_spec] expression to match the arguments

[match_spec]: http://erlang.org/doc/apps/erts/match_spec.html